+++ app/dashboard/Test/page.tsx
@@ -1,0 +1,11 @@
+
+export default function Test() {
+
+    return (
+        <>
+        <div>
+            <p>test page </p>
+        </div>
+        </>
+    )
+}
