+++ package.json
@@ -17,7 +17,8 @@
     "dotenv": "^16.3.1",
     "flowbite": "^2.2.0",
     "flowbite-react": "^0.7.2",
-    "next": "14.0.4",
+    "next": "^14.0.4",
+    "node": "^21.2.0",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0",
     "pg": "^8.11.3",
